move, remove index folders, better tree shaking
