+++ utils/base64.ts
@@ -1,5 +1,9 @@
 type Base64String = string;
 
+export default function btoa(str: string): Base64String {
+  return Buffer.from(str, "base64").toString();
+}
+
 export default function atob(str: string | Buffer): Base64String {
   if (str instanceof Buffer) {
     return str.toString("base64");
