--- utils/icons/index.d.ts
@@ -1,4 +1,0 @@
-export as namespace icons;
-
-export function downloadIcons(storageId: number, manifest: any): void;
-export function isDataUri(uri: string): boolean;
