--- utils/storage/index.d.ts
@@ -1,3 +1,0 @@
-export as namespace storage;
-
-export function createId(siteUrl: string): string;
